Create Expect.java
This is a slightly modified version of Ronnie Dong's Expect for Java.  The original can be found here: https://github.com/ronniedong/Expect-for-Java

The modifications remove most logging because it was originally written for an old version of log4j, and for my purposes the logging wasn't necessary.  I updated some of the code for that with the idea that if, at some point in the future, I get around to updating the whole thing, I'll issue a pull request or fork it or whatnot to provide an updated version.  As it is in the original repo, it doesn't work.  So I just changed it so it would work for this project.